feat(readme_writer): add template-based README generation
Added an option to generate README files based on a template. This new feature allows for more flexibility in the structure and content of generated READMEs. The function `write_readme` now accepts an additional argument `use_template`, which when set to True, generates a templated README. If False, it follows the previous method of generating a list-based README.